knowledge_extractor/pyproject.toml
[project]
name = "knowledge-extractor"
version = "0.1.0"
description = "CLI tool for generating YAML knowledge bases from unstructured documents"
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    # LLM
    "openai>=1.0.0",
    "outlines>=0.1.0",
    "pydantic>=2.0.0",

    # Document parsing
    "pymupdf>=1.23.0",
    "python-docx>=1.0.0",
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "chardet>=5.0.0",

    # NLP
    "pymorphy3>=1.2.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",

    # Utils
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
knowledge-extractor = "knowledge_extractor.__main__:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py310"
